email conformation and password reset more code
